Lots of refactoring of javascript.  Could still use a lot more.
