+++ templates/account/signup.html
@@ -3,7 +3,7 @@
 {% load i18n %}
 {% load crispy_forms_tags %}
 
-{% block head_title %}SoftwareEngineering.Life: {% trans "Signup" %}{% endblock %}
+{% block head_title %}DRF React{% trans "Signup" %}{% endblock %}
 
 {% block form_content %}
     <h1>{% trans "Sign Up" %}</h1>
