+++ templates/base.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">    <title>SoftwareEngineering.life</title>
-    <title>{% block head_title %}SoftwareEngineering.Life{% endblock %}</title>
+    <title>{% block head_title %}DRF React{% endblock %}</title>
 </head>
 <body>
 <section class="section">
